Migrado o calculo de datas para uma classe do util
