+++ src/main/java/com/alucarweb/devolution/DevolutionController.java
@@ -1,7 +1,5 @@
 package com.alucarweb.devolution;
 
-import java.util.Calendar;
-
 import javax.inject.Inject;
 
 import com.alucarweb.annotation.TransactionRequired;
@@ -11,6 +9,7 @@ import com.alucarweb.rent.RentController;
 import com.alucarweb.rent.RentDAO;
 import com.alucarweb.util.AlucarConfig;
 import com.alucarweb.util.AlucarConfig.Property;
+import com.alucarweb.util.DateCalc;
 
 import br.com.caelum.vraptor.Controller;
 import br.com.caelum.vraptor.Post;
@@ -36,7 +35,7 @@ public class DevolutionController {
 		
 		//CALCULO DO VALOR BASEADO NOS DIAS
 		double totalValue = 0;
-		int diasLocacao = daysBetween(rent.getExpectedDate(),rent.getCreatedAt());
+		int diasLocacao = DateCalc.daysBetween(rent.getExpectedDate(),rent.getCreatedAt());
 		if(rent.getKillometerType() == KillometerType.CONTROLLED){
 			totalValue += Double.parseDouble(rent.getCar().getControlKm()) * diasLocacao;
 		}else{
@@ -44,7 +43,7 @@ public class DevolutionController {
 		}
 		
 		//MULTA POR ATRASO DE DIA BASEADO NO VALOR DA LOCAÇÃO
-		int diasAtraso = daysBetween(rent.getExpectedDate(),devolution.getDate()); 
+		int diasAtraso = DateCalc.daysBetween(rent.getExpectedDate(),devolution.getDate()); 
 		if(diasAtraso > 0){
 			diasAtraso = (diasAtraso > 20) ? 20 : diasAtraso;
 			double valorAtraso = ((diasAtraso / 100) * totalValue);
@@ -78,28 +77,5 @@ public class DevolutionController {
 	}
 	
 	
-	public static int daysBetween(Calendar day1, Calendar day2){
-	    Calendar dayOne = (Calendar) day1.clone();
-	    Calendar dayTwo = (Calendar) day2.clone();
-
-	    if (dayOne.get(Calendar.YEAR) == dayTwo.get(Calendar.YEAR)) {
-	        return Math.abs(dayOne.get(Calendar.DAY_OF_YEAR) - dayTwo.get(Calendar.DAY_OF_YEAR));
-	    }
-        
-	    if (dayTwo.get(Calendar.YEAR) > dayOne.get(Calendar.YEAR)) {
-            Calendar temp = dayOne;
-            dayOne = dayTwo;
-            dayTwo = temp;
-        }
-        int extraDays = 0;
-
-        int dayOneOriginalYearDays = dayOne.get(Calendar.DAY_OF_YEAR);
-
-        while (dayOne.get(Calendar.YEAR) > dayTwo.get(Calendar.YEAR)) {
-            dayOne.add(Calendar.YEAR, -1);            // 
-            extraDays += dayOne.getActualMaximum(Calendar.DAY_OF_YEAR);
-        }
-
-        return extraDays - dayTwo.get(Calendar.DAY_OF_YEAR) + dayOneOriginalYearDays ;
-	}
+	
 }
